pyproject.toml
[project]
name = "funboost-demo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "funboost[all]",
    "python-dotenv>=1.2.2",
    "redis[cluster]",
    "uvicorn>=0.44.0",
]
